BackEnd/requirements.txt
alembic==1.4.2
Babel==2.8.0
click==7.1.2
ecdsa==0.15
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
phonenumbers==8.12.6
psycopg2-binary==2.8.5
pyasn1==0.4.8
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2020.1
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.1
